+++ central/Cargo.toml
@@ -15,3 +15,4 @@ futures = { workspace = true }
 serde_json = "1"
 rand = "0.8"
 reqwest = { version = "0.12", default_features = false, features = ["default-tls"] }
+anyhow = "1.0.91"
