+++ central/src/auth/authentik/test.rs
@@ -2,9 +2,9 @@ use beam_lib::reqwest::{self, Error, StatusCode, Url};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use shared::{OIDCConfig, SecretResult};
-use crate::auth::authentik::app::{combine_application, generate_app_values};
+use crate::auth::authentik::app::generate_app_values;
 use crate::auth::authentik::group::create_groups;
-use crate::auth::authentik::{app_configs_match, compare_applications, get_application, get_uuid, AuthentikConfig};
+use crate::auth::authentik::{app_configs_match, combine_app_provider, compare_applications, get_application, get_uuid, AuthentikConfig};
 use crate::CLIENT;
 
 #[derive(Deserialize, Serialize, Debug)]
@@ -79,33 +79,33 @@ async fn setup_authentik() -> reqwest::Result<(String, AuthentikConfig)> {
 
 #[ignore = "Requires setting up a authentik"]
 #[tokio::test]
-async fn test_create_client() -> reqwest::Result<()> {
+async fn test_create_client() -> anyhow::Result<()> {
     let (token, conf) = setup_authentik().await?;
     let name = "home";
     // public client
     let client_config = OIDCConfig { is_public: true, redirect_urls: vec!["http://foo/bar".into()] };
-    let (SecretResult::Created(pw) | SecretResult::AlreadyExisted(pw)) = dbg!(combine_application(&token, name, &client_config, &conf).await?) else {
+    let (SecretResult::Created(pw) | SecretResult::AlreadyExisted(pw)) = dbg!(combine_app_provider(&token, name, &client_config, &conf).await?) else {
         panic!("Not created or existed")
     };
     let c = dbg!(get_application(name, &token, &client_config, &conf).await.unwrap());
-    assert!(app_configs_match(&c, &generate_app_values(name, name, &client_config, &pw)));
-    assert!(dbg!(compare_applications(&token, name, &client_config, &conf, &pw).await?));
+    assert!(app_configs_match(&c, &generate_app_values(name, name, &client_config)));
+    assert!(dbg!(compare_applications(&token, name, &client_config, &conf).await?));
 
     // private client
     let client_config = OIDCConfig { is_public: false, redirect_urls: vec!["http://foo/bar".into()] };
-    let (SecretResult::Created(pw) | SecretResult::AlreadyExisted(pw)) = dbg!(combine_application(&token, name, &client_config, &conf).await?) else {
+    let (SecretResult::Created(pw) | SecretResult::AlreadyExisted(pw)) = dbg!(combine_app_provider(&token, name, &client_config, &conf).await?) else {
         panic!("Not created or existed")
     };
     let c = dbg!(get_application(name, &token, &client_config, &conf).await.unwrap());
-    assert!(app_configs_match(&c, &generate_app_values(name, name, &client_config, &pw)));
-    assert!(dbg!(compare_applications(&token, name, &client_config, &conf, &pw).await?));
+    assert!(app_configs_match(&c, &generate_app_values(name, name, &client_config)));
+    assert!(dbg!(compare_applications(&token, name, &client_config, &conf).await?));
 
     Ok(())
 }
 
 #[ignore = "Requires setting up a authentik"]
 #[tokio::test]
-async fn group_test() -> reqwest::Result<()> {
+async fn group_test() -> anyhow::Result<()> {
     let (token, conf) = setup_authentik().await?;
     create_groups("e", &token, &conf).await?;
     // dbg!(get_realm_permission_roles(&token, &conf).await?);
@@ -120,9 +120,6 @@ async fn test_flow_property() {
     let test_key = "groups";
     let flow_url = "/api/v3/flows/instances/?ordering=slug&page=1&page_size=20&search=";
     let res = get_uuid(flow_url, &conf, &token, test_key).await;
-    if res.is_empty() {
-    } else {
-        dbg!(res);    
-    }
+        dbg!(res);
     
 }
