+++ central/src/auth/mod.rs
@@ -1,3 +1,3 @@
 mod keycloak;
 pub(crate) mod config;
-mod authentik;
+pub mod authentik;
